native: refactoring, move to anonymous namespace
